fix error type, change Long to Integer

+++ src/main/java/cu/edu/mes/sigenu/training/core/repository/QuestionRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 import cu.edu.mes.sigenu.training.core.model.Question;
 
-public interface QuestionRepository extends JpaRepository<Question, Long>, JpaSpecificationExecutor<Question> {
+public interface QuestionRepository extends JpaRepository<Question, Integer>, JpaSpecificationExecutor<Question> {
 	
 	Question findByQuestion(String question);
 }
